refactor: 인수테스트와 컨트롤러 테스트를 분리
