+++ README.md
@@ -25,10 +25,15 @@
 
 ## 1단게 수정 사항 & 피드백 정리
 
-- [ ] Acceptance test를 하면서 실제 사용자가 생각하는 흐름을 정리해보자
+- [x] Acceptance test를 하면서 실제 사용자가 생각하는 흐름을 정리해보자
   - 위의 두가지 경우처럼 나눌 수도 있고 더 다양하게 나누는 방법이 있을텐데요.
     알린이 생각하는 시나리오를 한번 생각해보고 그것에 맞게 테스트를 작성해 보면 좋을 것 같아요!
     시나리오로 모든 경우를 테스트할 수 없기 때문에 controller로 통합테스트를 진행하는 것이지 않을까요?
+  - 이전까지 인수 테스트에 대한 감이 잘 안와서 모든 E2E 테스트를 인수테스트에서 한 것 같습니다.
+  - 피드백에 대해 생각해보다 보니 인수테스트에서는 사용자 입장에서의 시나리오를 작성하여 테스트하고 컨트롤러 테스트에서는 기능 테스트와 예외에 대한 테스트를 하도록 분리할 수 있을 것 같았습니다.
+  - 인수 테스트를 할 때 제가 생각하는 시나리오로 인수테스트를 할 수 있도록 수정해보고 주석으로 시나리오를 써봤습니다.
+  - 예외 테스트 등은 컨트롤러 테스트로 옮겼습니다.
+  - 컨트롤러 테스트에서는 DirtiesContext를 하지 않으면서 속도가 매우 빨라졌습니다.
 - [ ] interceptor가 ui 패키지에 포함되어 있다. ui보다 config에 있는게 맞지 않을까? ui에 포함하게된 기준이 무엇이었을까?
 - [ ] webpage를 보며 me라는 표현을 본적이 없지 않나요?
 - [x] exception과 runtimeException을 같이 잡은 이유가 있을까요??
