+++ src/test/java/woowacourse/auth/acceptance/AuthAcceptanceTest.java
@@ -1,9 +1,12 @@
 package woowacourse.auth.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -17,11 +20,14 @@ import woowacourse.shoppingcart.dto.SignUpRequest;
 @DisplayName("인증 관련 기능")
 public class AuthAcceptanceTest extends AcceptanceTest {
 
-    @DisplayName("Bearer 토큰이 생성된다")
+    @DisplayName("회원가입을 하고 로그인을 하면 토큰을 반환한다.")
     @Test
     void createToken() {
-        // given
-        SignUpRequest signUpRequest = new SignUpRequest("rennon", "rennon@woowa.com", "12345678");
+        // given(회원가입을 하고)
+        String username = "rennon";
+        String email = "rennon@woowa.com";
+        String password = "12345678";
+        SignUpRequest signUpRequest = new SignUpRequest(username, email, password);
         RestAssured
                 .given().log().all()
                 .contentType(ContentType.JSON)
@@ -29,26 +35,30 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                 .when().post("/users")
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value());
-        SignInRequest signInRequest = new SignInRequest("rennon@woowa.com", "12345678");
 
-        // when
-        String token = RestAssured
+        // when(등록된 email, password로 로그인을 하면)
+        SignInRequest signInRequest = new SignInRequest(email, password);
+        ExtractableResponse<Response> response = RestAssured
                 .given().log().all()
                 .body(signInRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().post("/login")
-                .then().log().all().extract().as(SignInResponse.class).getToken();
+                .then().log().all().extract();
 
-        // then
+        // then(토큰을 반환한다.)
+        String token = response.as(SignInResponse.class).getToken();
         assertThat(token).isNotBlank();
     }
 
-    @DisplayName("Bearer Auth 로그인 성공")
+    @DisplayName("")
     @Test
     void myInfoWithBearerAuth() {
-        // given
-        SignUpRequest signUpRequest = new SignUpRequest("rennon", "rennon@woowa.com", "12345678");
+        // given(회원가입하고 로그인을 하고)
+        String username = "rennon";
+        String email = "rennon@woowa.com";
+        String password = "12345678";
+        SignUpRequest signUpRequest = new SignUpRequest(username, email, password);
         RestAssured
                 .given().log().all()
                 .contentType(ContentType.JSON)
@@ -57,7 +67,7 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value());
 
-        SignInRequest signInRequest = new SignInRequest("rennon@woowa.com", "12345678");
+        SignInRequest signInRequest = new SignInRequest(email, password);
         String token = RestAssured
                 .given().log().all()
                 .body(signInRequest)
@@ -66,27 +76,32 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                 .when().post("/login")
                 .then().log().all().extract().as(SignInResponse.class).getToken();
 
-        // when
-        // 발급 받은 토큰을 사용하여 내 정보 조회를 요청하면
-        CustomerResponse response = RestAssured
+        // when(발급 받은 토큰을 사용하여 내 정보 조회를 요청하면)
+        ExtractableResponse<Response> response = RestAssured
                 .given().log().all()
                 .header("Authorization", "Bearer " + token)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/users/me")
                 .then().log().all()
-                .statusCode(HttpStatus.OK.value()).extract().as(CustomerResponse.class);
-        // then
-        // 내 정보가 조회된다
+                .extract();
 
-        assertThat(response.getEmail()).isEqualTo("rennon@woowa.com");
+        // then(내 정보가 조회된다)
+        CustomerResponse customerResponse = response.as(CustomerResponse.class);
+        assertAll(
+                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
+                () -> assertThat(customerResponse.getUsername()).isEqualTo(username),
+                () -> assertThat(customerResponse.getEmail()).isEqualTo(email)
+        );
     }
 
     @DisplayName("Bearer Auth 로그인 실패")
     @Test
     void myInfoWithBadBearerAuth() {
-        // given
-        // 회원이 등록되어 있고
-        SignUpRequest signUpRequest = new SignUpRequest("rennon", "rennon@woowa.com", "12345678");
+        // given(회원가입을 하고)
+        String username = "rennon";
+        String email = "rennon@woowa.com";
+        String password = "12345678";
+        SignUpRequest signUpRequest = new SignUpRequest(username, email, password);
         RestAssured
                 .given().log().all()
                 .contentType(ContentType.JSON)
@@ -95,26 +110,29 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value());
 
-        // when
-        // 잘못된 id, password를 사용해 토큰을 요청하면
-        // then
-        // 토큰 발급 요청이 거부된다
-        SignInRequest signInRequest = new SignInRequest("rennon@woowa.com", "56781234");
-        RestAssured
+        // when(잘못된 email, password를 사용해 토큰을 요청하면)
+        String wrongEmail = "greenlawn@woowa.com";
+        SignInRequest signInRequest = new SignInRequest(wrongEmail, password);
+        ExtractableResponse<Response> response = RestAssured
                 .given().log().all()
                 .body(signInRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().post("/login")
-                .then().log().all()
-                .statusCode(HttpStatus.BAD_REQUEST.value());
+                .then().log().all().extract();
+
+        // then(토큰 발급 요청이 거부된다)
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
     @DisplayName("Bearer Auth 유효하지 않은 토큰")
     @Test
     void myInfoWithWrongBearerAuth() {
-        // given
-        SignUpRequest signUpRequest = new SignUpRequest("rennon", "rennon@woowa.com", "12345678");
+        // given(회원가입하고 로그인을 하고)
+        String username = "rennon";
+        String email = "rennon@woowa.com";
+        String password = "12345678";
+        SignUpRequest signUpRequest = new SignUpRequest(username, email, password);
         RestAssured
                 .given().log().all()
                 .contentType(ContentType.JSON)
@@ -123,17 +141,26 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value());
 
-        // when
-        // 유효하지 않은 토큰을 사용하여 내 정보 조회를 요청하면
-        // then
-        // 내 정보 조회 요청이 거부된다
-        String token = "dummy";
-        RestAssured
+        SignInRequest signInRequest = new SignInRequest(email, password);
+        String token = RestAssured
                 .given().log().all()
-                .header("Authorization", "Bearer " + token)
+                .body(signInRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/login")
+                .then().log().all().extract().as(SignInResponse.class).getToken();
+
+        // when(유효하지 않은 토큰을 사용하여 내 정보 조회를 요청하면)
+        String wrongToken = "dummy";
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + wrongToken)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/users/me")
                 .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
+                .extract();
+
+        // then(내 정보 조회 요청이 거부된다)
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
     }
 }
